Rename inbound to internal and outbound to external
