+++ network/src/external/message/read.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::outbound::message::MessageHeader;
+use crate::external::message::MessageHeader;
 use snarkos_errors::network::message::{MessageError, MessageHeaderError, StreamReadError};
 
 use tokio::{io::AsyncRead, prelude::*};
