+++ network/src/external/message_types/sync.rs
@@ -14,41 +14,51 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::outbound::message::{Message, MessageName};
+use crate::external::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
+use snarkos_objects::BlockHeaderHash;
 
-#[cfg_attr(nightly, doc(include = "../../../documentation/network_messages/get_memory_pool.md"))]
+#[cfg_attr(nightly, doc(include = "../../../documentation/network_messages/sync.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct GetMemoryPool;
+pub struct Sync {
+    /// Known hashes of blocks to share
+    pub block_hashes: Vec<BlockHeaderHash>,
+}
 
-impl Message for GetMemoryPool {
+impl Sync {
+    pub fn new(block_hashes: Vec<BlockHeaderHash>) -> Self {
+        Self { block_hashes }
+    }
+}
+
+impl Message for Sync {
     fn name() -> MessageName {
-        MessageName::from("getmempool")
+        MessageName::from("sync")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
-        if vec.len() != 0 {
-            return Err(MessageError::InvalidLength(vec.len(), 0));
-        }
-
-        Ok(Self)
+        Ok(Self {
+            block_hashes: bincode::deserialize(&vec)?,
+        })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        Ok(vec![])
+        Ok(bincode::serialize(&self.block_hashes)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use snarkos_testing::consensus::BLOCK_1_HEADER_HASH;
 
     #[test]
-    fn test_get_memory_pool() {
-        let message = GetMemoryPool;
+    fn test_sync() {
+        let data = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
+        let message = Sync::new(vec![data]);
 
         let serialized = message.serialize().unwrap();
-        let deserialized = GetMemoryPool::deserialize(serialized).unwrap();
+        let deserialized = Sync::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
