+++ network/src/external/message_types/transaction.rs
@@ -14,51 +14,49 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::outbound::message::{Message, MessageName};
+use crate::external::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
-use snarkos_objects::BlockHeaderHash;
 
-#[cfg_attr(nightly, doc(include = "../../../documentation/network_messages/get_block.md"))]
+#[cfg_attr(nightly, doc(include = "../../../documentation/network_messages/transaction.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct GetBlock {
-    /// Header hash of requested block
-    pub block_hash: BlockHeaderHash,
+pub struct Transaction {
+    /// Serialized transaction bytes
+    pub(crate) bytes: Vec<u8>,
 }
 
-impl GetBlock {
-    pub fn new(block_hash: BlockHeaderHash) -> Self {
-        Self { block_hash }
+impl Transaction {
+    pub fn new(bytes: Vec<u8>) -> Self {
+        Self { bytes }
     }
 }
 
-impl Message for GetBlock {
+impl Message for Transaction {
     fn name() -> MessageName {
-        MessageName::from("getblock")
+        MessageName::from("transaction")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
         Ok(Self {
-            block_hash: bincode::deserialize(&vec)?,
+            bytes: bincode::deserialize(&vec)?,
         })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        Ok(bincode::serialize(&self.block_hash)?)
+        Ok(bincode::serialize(&self.bytes)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use snarkos_testing::consensus::BLOCK_1_HEADER_HASH;
+    use snarkos_testing::consensus::TRANSACTION_1;
 
     #[test]
-    fn test_block() {
-        let block_hash = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
-        let message = GetBlock::new(block_hash);
+    fn test_transaction() {
+        let message = Transaction::new(TRANSACTION_1.to_vec());
 
         let serialized = message.serialize().unwrap();
-        let deserialized = GetBlock::deserialize(serialized).unwrap();
+        let deserialized = Transaction::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
