+++ network/src/external/mod.rs
