+++ network/src/external/protocol/handshake/handshakes.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::outbound::{
+use crate::external::{
     message_types::{GetPeers, Verack, Version},
     Channel,
     Handshake,
@@ -178,7 +178,7 @@ impl Handshakes {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::outbound::Message;
+    use crate::external::Message;
     use snarkos_testing::network::random_socket_address;
 
     use serial_test::serial;
