+++ network/src/external/protocol/sync/sync.rs
@@ -15,8 +15,8 @@
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
+    external::Channel,
     message_types::{GetBlock, GetSync},
-    outbound::Channel,
 };
 use snarkos_errors::network::SendError;
 use snarkos_models::{algorithms::LoadableMerkleParameters, objects::Transaction};
