+++ network/src/server/context/connections.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::outbound::Channel;
+use crate::external::Channel;
 
 use std::{collections::HashMap, net::SocketAddr, sync::Arc};
